Refactor: Mutation 훅 리팩토링 및 api 호출 통합
