--- src/entities/group/model/use-create-group-info.ts
@@ -1,9 +1,0 @@
-import { useMutation } from "@tanstack/react-query";
-import { CreateGroupInfoRequest } from "./types";
-import { createGroupInfo } from "../api/create-group-info";
-
-export const useCreateGroupInfo = () => {
-    return useMutation({
-        mutationFn: (data: CreateGroupInfoRequest) => createGroupInfo(data)
-    });
-};
