--- src/entities/schdule/model/use-create-schedule.ts
@@ -1,9 +1,0 @@
-import { useMutation } from "@tanstack/react-query";
-import { CreateScheduleRequest } from "./types";
-import { createSchedule } from "../api/create-schedule";
-
-export const useCreateSchedule = () => {
-    return useMutation({
-        mutationFn: (data: CreateScheduleRequest) => createSchedule(data),
-    });
-};
