+++ src/features/create-group/api/use-create-group.ts
@@ -1,0 +1,26 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { post } from "@/features/auth/model/auth-client";
+import { CreateGroupRequest, GroupDetails } from "@/entities/group/model/types";
+import { groupQuery } from "@/entities/group/api/group.query";
+
+export const useCreateGroup = () => {
+    const queryClient = useQueryClient();
+
+    return useMutation({
+        mutationFn: async (data: CreateGroupRequest): Promise<GroupDetails> => {
+            const response = await post("/api/v1/groups", data);
+            const result = await response.json();
+
+            if (result.code !== 201) {
+                throw new Error(result.message || "모임 생성에 실패했습니다.");
+            }
+
+            return result.data;
+        },
+        onSuccess: () => {
+            queryClient.invalidateQueries({
+                queryKey: groupQuery.lists()
+            });
+        },
+    });
+};
